+++ src/packages/piral-cli/src/injectors/pilet.test.ts
@@ -1,7 +1,32 @@
-import PiletInjector from './piral';
+import PiletInjector from './pilet';
+import { EventEmitter } from 'events';
 
 describe('Piral-CLI piral injector', () => {
-  it('PiletInjector communicates via WebSocket', () => {
-
+  it('PiletInjector is active when configured', () => {
+    const bundler = {
+      pending: true,
+      bundle: {
+        dir: '',
+        hash: '',
+        name: '',
+      },
+      on() {},
+      off() {},
+      start() {},
+      ready() {
+        return Promise.resolve();
+      },
+    };
+    const config = {
+      bundler,
+      root: '',
+      port: 1234,
+      api: '',
+      app: '',
+      active: true,
+    };
+    const core = new EventEmitter();
+    const injector = new PiletInjector(config, undefined, core);
+    expect(injector.active).toBeTruthy();
   });
 });