+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Added optional `--cacert` flag to `pilet publish` command (#204)
 * Fixed `NODE_ENV` mode of the emulator (#207)
 * Added `getProfile()` to [piral-oidc](./src/packages/piral-oidc/README.md) (#210)
+* Switched to relative paths for local Piral references
+* Added new package `siteless` for painless microfrontends
 
 ## 0.11.3 (April 24, 2020)